Separate the app level bar layout tests from the update banner's
Two tests covered how the bars sit together rather than anything about the
update banner they happened to live with. They move to their own spec, waiting
for the banner by polling because a single fast forward can be swallowed.
