+++ playwright/CLAUDE.md
@@ -84,7 +84,7 @@ Helpers log in as `admin:test` internally where admin rights are needed. Test cr
 
 ## Tests (`playwright/tests/`)
 
-Specs cover the main user flows: auth (`login`, `kompassiLogin`, `registration`, `logout`), sign-up (`lotterySignup`, `directSignup`, `preConventionWeekSignup`, `favorite`, `admission`, `myProgramItems` — the My Program lists' cancel/unfavorite actions, empty states, and the persisted upcoming/all starting-time toggle), groups (`group`, `groupDirectSignup`), discovery (`programSearch`, `programTagFilter` — the multi-select tag filter's OR logic and sessionStorage persistence, `programListFilters` — the hide-full-items checkbox, the revolving door starting-time view, and persistence of all the filter card's selections (program type, starting time, search term, hide-full) over a reload, `programListQueryParams` — the `programType`/`invalid` program-list URL params, `programItemErrors` — the per-card error messages for items missing required info (attendance limits, sign-up type), `programListScroll` — the virtualized list's scroll restore/reset on navigation, `programDateDisplay` — program times show full dates outside event week and weekday-only during it), staff tools (`admin`, `adminConsole`, `helper`), `profile`, `about`, `dashboard` — the public `/dashboard` lottery-results view (no login required), program-item lifecycle (`cancelProgramItem`, `moveProgramItem` — the cancellation/start-time-change notifications described in [server/CLAUDE.md](../server/CLAUDE.md)), `dataPolling` — the app's periodic background data refresh (`dataUpdateInterval`, 60 s) delivering server-side changes without navigation, `appUpdateBanner` - the new-version banner (patches the polled settings response to fake the server-reported version: confirmation-window acceptance, dismiss, the reload button, and the transparent once-per-version reload on the first navigation), and `networkError` — the network-error toast state machine (toast on outage after the decision delay, healing on the next successful response, grace-period suppression of a failing refresh right after reconnect, the probe surfacing a genuine outage, HTTP errors on background requests counting as connectivity issues inside the grace period and as API errors outside it, immediate toasts for user-initiated failures, and the profile email form reporting a failed follow-up session refresh instead of success). These use `page.clock.install()` + `fastForward()` to trigger timers instantly — install the mocked clock **before** `page.goto` so the app's `setInterval` uses it; promise/fetch resolution is unaffected, only timers. `networkError` simulates network failure with `page.route(...)` + `route.abort()` (not `context.setOffline`, which would flip `navigator.onLine` and route the client into its offline branch), pauses the mocked clock after load (`clock.pauseAt` — an installed clock otherwise keeps advancing in real time, letting slow CI steps erode the mocked grace period), advances the clock in small steps so fetch rejections can settle and schedule their follow-up timers between jumps, and dispatches a synthetic `online` event to stamp the reconnect grace period (the page-resume listeners themselves aren't drivable — Playwright has no visibility emulation).
+Specs cover the main user flows: auth (`login`, `kompassiLogin`, `registration`, `logout`), sign-up (`lotterySignup`, `directSignup`, `preConventionWeekSignup`, `favorite`, `admission`, `myProgramItems` — the My Program lists' cancel/unfavorite actions, empty states, and the persisted upcoming/all starting-time toggle), groups (`group`, `groupDirectSignup`), discovery (`programSearch`, `programTagFilter` — the multi-select tag filter's OR logic and sessionStorage persistence, `programListFilters` — the hide-full-items checkbox, the revolving door starting-time view, and persistence of all the filter card's selections (program type, starting time, search term, hide-full) over a reload, `programListQueryParams` — the `programType`/`invalid` program-list URL params, `programItemErrors` — the per-card error messages for items missing required info (attendance limits, sign-up type), `programListScroll` — the virtualized list's scroll restore/reset on navigation, `programDateDisplay` — program times show full dates outside event week and weekday-only during it), staff tools (`admin`, `adminConsole`, `helper`), `profile`, `about`, `dashboard` — the public `/dashboard` lottery-results view (no login required), program-item lifecycle (`cancelProgramItem`, `moveProgramItem` — the cancellation/start-time-change notifications described in [server/CLAUDE.md](../server/CLAUDE.md)), `dataPolling` — the app's periodic background data refresh (`dataUpdateInterval`, 60 s) delivering server-side changes without navigation, `appUpdateBanner` - the new-version banner (patches the polled settings response to fake the server-reported version: confirmation-window acceptance, dismiss, the reload button, and the transparent once-per-version reload on the first navigation), `appLevelBars` — how the bars below the header sit together rather than what any one of them does (sticky stacking without overlap, and the dismiss icons sharing a right edge and vertical centre); each bar's own behaviour is covered where its feature lives, and `networkError` — the network-error toast state machine (toast on outage after the decision delay, healing on the next successful response, grace-period suppression of a failing refresh right after reconnect, the probe surfacing a genuine outage, HTTP errors on background requests counting as connectivity issues inside the grace period and as API errors outside it, immediate toasts for user-initiated failures, and the profile email form reporting a failed follow-up session refresh instead of success). These use `page.clock.install()` + `fastForward()` to trigger timers instantly — install the mocked clock **before** `page.goto` so the app's `setInterval` uses it; promise/fetch resolution is unaffected, only timers. `networkError` simulates network failure with `page.route(...)` + `route.abort()` (not `context.setOffline`, which would flip `navigator.onLine` and route the client into its offline branch), pauses the mocked clock after load (`clock.pauseAt` — an installed clock otherwise keeps advancing in real time, letting slow CI steps erode the mocked grace period), advances the clock in small steps so fetch rejections can settle and schedule their follow-up timers between jumps, and dispatches a synthetic `online` event to stamp the reconnect grace period (the page-resume listeners themselves aren't drivable — Playwright has no visibility emulation).
 
 ### Page Objects (`pages/`)
